refactored data type in empirical distribution

+++ statistics/empiricalDistribution_test.go
@@ -6,8 +6,8 @@ import (
 )
 
 func TestEmpiricalDistribution_GetSampleSize(t *testing.T) {
-	binstarts := []int64{0, 1, 2, 3, 4}
-	bincounts := []int64{2, 3, 4, 3, 2}
+	binstarts := []float64{0, 1, 2, 3, 4}
+	bincounts := []float64{2, 3, 4, 3, 2}
 	e, _ := Init(binstarts, bincounts)
 	var expected int64 = 14
 	got := e.GetSampleSize()
@@ -16,8 +16,8 @@ func TestEmpiricalDistribution_GetSampleSize(t *testing.T) {
 	}
 }
 func TestEmpiricalDistribution_InvCDF(t *testing.T) {
-	binstarts := []int64{0, 1, 2, 3, 4}
-	bincounts := []int64{2, 3, 4, 3, 2}
+	binstarts := []float64{0, 1, 2, 3, 4}
+	bincounts := []float64{2, 3, 4, 3, 2}
 	e, _ := Init(binstarts, bincounts)
 	var expected float64 = 2
 	got := e.InvCDF(0.4)
@@ -27,8 +27,8 @@ func TestEmpiricalDistribution_InvCDF(t *testing.T) {
 	}
 }
 func TestEmpiricalDistribution_CDF(t *testing.T) {
-	binstarts := []int64{0, 1, 2, 3, 4}
-	bincounts := []int64{2, 3, 4, 3, 2}
+	binstarts := []float64{0, 1, 2, 3, 4}
+	bincounts := []float64{2, 3, 4, 3, 2}
 	e, _ := Init(binstarts, bincounts)
 	expected := 0.642857
 	got := e.CDF(3.5)
@@ -39,8 +39,8 @@ func TestEmpiricalDistribution_CDF(t *testing.T) {
 }
 
 func TestEmpiricalDistribution_PDF(t *testing.T) {
-	binstarts := []int64{0, 1, 2, 3, 4}
-	bincounts := []int64{2, 3, 4, 3, 2}
+	binstarts := []float64{0, 1, 2, 3, 4}
+	bincounts := []float64{2, 3, 4, 3, 2}
 	e, _ := Init(binstarts, bincounts)
 	var expected float64 = 0.21428571
 	got := e.PDF(3.5)
